📦 chore: 스토리북에서 GlobalStyle과 theme를 사용할 수 있도록 설정
