+++ .storybook/preview.tsx
@@ -1,0 +1,30 @@
+import type { Preview } from '@storybook/react';
+import { ThemeProvider } from 'styled-components';
+import GlobalStyle from '../src/styles/GlobalStyle';
+import { theme } from '../src/styles/theme';
+import React from 'react';
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+};
+
+export const decorators = [
+  (Story) => (
+    <>
+      <ThemeProvider theme={theme}>
+        <GlobalStyle />
+        <Story />
+      </ThemeProvider>
+    </>
+  ),
+];
+
+export default preview;
